app/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
openai
langchain
pydantic
spacy
nltk
pdfminer.six